add getSiteMapUrls utility
